Grid + Logic Loop
Solved the issue of looping and timings with the use of QTimer()

Semi-implementation of Hans grid system

Looked into animations but haven't implemented anything.

Planned for the implementation of interrupts/error fixing in the programm by creating a very simple state machine.

Moved all timer logic and individual task rendering components to the task file itself.